perf: add memoization on editor hooks
